bump to 16, most frontend working

+++ OpenAlprWebhookProcessor/ClientApp/src/app/users/list.component.ts
@@ -1,11 +1,12 @@
 ﻿import { Component, OnInit } from '@angular/core';
 import { first } from 'rxjs/operators';
 
-import { AccountService } from '@app/_services';
+import { AccountService } from 'app/_services';
+import { User } from 'app/_models';
 
 @Component({ templateUrl: 'list.component.html' })
 export class ListComponent implements OnInit {
-    users = null;
+    users: User[] = [];
 
     constructor(private accountService: AccountService) {}
 
@@ -17,9 +18,11 @@ export class ListComponent implements OnInit {
 
     deleteUser(id: string) {
         const user = this.users.find(x => x.id === id);
-        user.isDeleting = true;
-        this.accountService.delete(id)
-            .pipe(first())
-            .subscribe(() => this.users = this.users.filter(x => x.id !== id));
+        if(user !== undefined) {
+            user.isDeleting = true;
+            this.accountService.delete(id)
+                .pipe(first())
+                .subscribe(() => this.users = this.users.filter(x => x.id !== id));
+        }
     }
 }
